feat(BE): first draft for a BE setup (WIP)
