+++ be/src/types/action.ts
@@ -1,0 +1,13 @@
+/**
+ * Interface for any Action (or message) sent between Client and Server via Websocket
+ * @param {string} type defines the type of the action, e.g connecting, drawing, erasing, set_background
+ * @param {Object} data any data that belongs to the action
+ * @param {Date} timestamp
+ * @param {string} userId unique identifier for each client / user
+ */
+export class Action {
+    type: string
+    data?: Object
+    timestamp?: Date
+    userId?: string
+}
