+++ be/src/utils/utils.ts
@@ -1,0 +1,7 @@
+export const uuidv4 = () =>
+  ('' + 1e7 + -1e3 + -4e3 + -8e3 + -1e11).replace(/[018]/g, c =>
+    (
+      +c ^
+      ((window.crypto || (window as any).msCrypto).getRandomValues(new Uint8Array(1))[0]! & (15 >> (+c / 4)))
+    ).toString(16),
+  );
