+++ be/tsconfig.json
@@ -1,0 +1,14 @@
+{
+    "compilerOptions": {                        
+      "target": "es6",                               
+      "module": "commonjs",                           
+      "outDir": "./dist",                             
+      "rootDir": "",                             
+      "strict": false,
+      "moduleResolution": "node",
+      "esModuleInterop": true,                       
+    },
+    "exclude":[
+      "./node_modules"
+    ]
+  }
